+++ src/Markdig.SyntaxHighlighting.Tests/Markdig.SyntaxHighlighting.Tests.csproj
@@ -10,6 +10,12 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <Compile Remove="Properties\**" />
+    <EmbeddedResource Remove="Properties\**" />
+    <None Remove="Properties\**" />
+  </ItemGroup>
+
+  <ItemGroup>
     <Compile Remove="Example\CodeSample.cs" />
   </ItemGroup>
 
@@ -53,11 +59,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\Markdig.SyntaxHighlighting\MarkdownServer.Markdig.SyntaxHighlighting.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Properties\" />
+    <ProjectReference Include="..\MarkdownServer.Markdig.SyntaxHighlighting\MarkdownServer.Markdig.SyntaxHighlighting.csproj" />
   </ItemGroup>
 
 </Project>
